Изменения: 1)size и data_capacity теперь имеют тип size_t 2)функция create_empty_dynamic_array использует внутри себя create_dynamic_array 3)вывод ошибок сделан через perror, чтобы ошибки выводились текстом, а не в виде кодов 4)map, where, concat не выделяют память внутри, а принимают 3 параметром, куда сохранять результат

+++ headers/FieldInfo.h
@@ -2,13 +2,15 @@
 #define FIRSTLAB_FIELD_INFO_H
 
 #include <stdbool.h>
+#include <stddef.h>
+
 #include "Allocator.h"
 
 typedef bool (*UnaryOperator)(void *, const void *);
 typedef int (*CompareOperator)(const void *, const void *);
 
 typedef struct {
-    int size;
+    size_t size;
     char *name;
     UnaryOperator assign;
     Allocator allocate;
